Updated the type of lagT and lagSW from double to long.

+++ crhmcode/src/modules/Classglacier.h
@@ -68,8 +68,8 @@ double* lagT_delayed{ NULL }; // lagged temperature, 22 Apr 2022 added
 double** lagT_delayed_lay{ NULL }; // lagged temperature, 22 Apr 2022 added
 double* lagSW_delayed{ NULL }; // lagged shortwave radiation, 22 Apr 2022 added
 double** lagSW_delayed_lay{ NULL }; // lagged shortwave radiation, 22 Apr 2022 added
-double* lagT{ NULL }; // debris-cover melt
-double* lagSW{ NULL }; // debris-cover melt
+long* lagT{ NULL }; // debris-cover melt
+long* lagSW{ NULL }; // debris-cover melt
 double* TF{ NULL }; // debris-cover melt
 double* SRF{ NULL }; // debris-cover melt
 double* Xdebris_melt_hrly{ NULL }; // debris-cover melt
